+++ __init__.py
@@ -105,7 +105,7 @@ from .validation import PLANETKA_OT_ReportBug, PLANETKA_OT_ValidateTextureSource
 bl_info = {
     "name": "Planetka - the Earth",
     "author": "Tomas Griger",
-    "version": (0, 3, 1),
+    "version": (0, 3, 2),
     "blender": (3, 6, 0),
     "location": "View3D > Sidebar > Planetka",
     "description": "Cinematic Earth visualisation system",